Installed svgr/webpack, Downloaded Login.svg, Deleted template and test-login pages, updated next.config.mjs and crea ted UI for login page (renamed signin to login page)
