+++ my-app/package.json
@@ -29,6 +29,7 @@
     "react-simple-maps": "^3.0.0"
   },
   "devDependencies": {
+    "@svgr/webpack": "^8.1.0",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
